cartItems 이벤트 리스너를 createCartItemsElement로 이동
